backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
PyMuPDF==1.24.3
spacy==3.7.4
sentence-transformers==2.7.0
scikit-learn==1.4.2
numpy==1.26.4
pydantic==2.7.1
SQLAlchemy==2.0.31
passlib==1.7.4
python-jose[cryptography]==3.3.0
